# server/spec/openapi/openapi.yaml
...
openapi: "3.0.3"
info:
  title: BenxinAdminPro API
  description: |
    本心通用管理后台底座 API 规范。
    v0.1.0: 安全头、C 端加密信封格式和统一安全错误码段。
    v0.2.0: 认证接口（登录/刷新/登出/验证码）。
    v0.3.0: 组织架构（用户/部门/岗位 CRUD）。
    v0.4.0: RBAC 核心（角色/菜单/权限 + Casbin 联动 + Hashid）。
  version: "0.4.0"
  contact:
    name: 仗键天涯(daxing)
    email: 3442535897@qq.com
servers:
  - url: http://localhost:8080
    description: 本地开发

# ---------------------------------------------------------------------------
# 安全定义
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0=成功，其它=错误码
        message:
          type: string
        data:
          description: 业务数据
...
    SysDept:
      type: object
      properties:
        id: { type: integer, format: int64 }
        parent_id: { type: integer, format: int64 }
        ancestors: { type: string }
        name: { type: string }
        sort: { type: integer }
        leader: { type: string }
        status: { type: integer }
        children:
          type: array
          items: { $ref: "#/components/schemas/SysDept" }
...
    SysRole:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        sort: { type: integer }
        status: { type: integer }
        remark: { type: string }
    SysMenu:
      type: object
      properties:
        id: { type: integer, format: int64 }
        parent_id: { type: integer, format: int64 }
        ancestors: { type: string }
        menu_type: { type: string, enum: [M, C, F] }
        name: { type: string }
        perm_code: { type: string }
        path: { type: string }
        component: { type: string }
        icon: { type: string }
        sort: { type: integer }
        visible: { type: integer }
        status: { type: integer }
        children:
          type: array
          items: { $ref: "#/components/schemas/SysMenu" }
...
paths:
...
  /sys/depts/tree:
    get:
      tags: [System]
      summary: 部门树
      operationId: getDeptTree
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: 部门树
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/SysDept" }
  /sys/depts:
    post:
      tags: [System]
      summary: 新建部门
      operationId: createDept
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysDept" }
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SysDept" }
  /sys/depts/{id}:
    put:
      tags: [System]
      summary: 更新部门
      operationId: updateDept
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysDept" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    delete:
      tags: [System]
      summary: 删除部门
      operationId: deleteDept
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
...
  /sys/roles:
    get:
      tags: [RBAC]
      summary: 角色列表
      operationId: listRoles
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    post:
      tags: [RBAC]
      summary: 新建角色
      operationId: createRole
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysRole" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/roles/{id}:
    get:
      tags: [RBAC]
      summary: 角色详情
      operationId: getRole
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    put:
      tags: [RBAC]
      summary: 更新角色
      operationId: updateRole
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysRole" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    delete:
      tags: [RBAC]
      summary: 删除角色
      operationId: deleteRole
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/roles/{id}/menus:
    put:
      tags: [RBAC]
      summary: 分配角色菜单/权限
      operationId: assignRoleMenus
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menu_ids: { type: array, items: { type: integer } }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/menus/tree:
    get:
      tags: [RBAC]
      summary: 菜单树
      operationId: getMenuTree
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/menus:
    post:
      tags: [RBAC]
      summary: 新建菜单
      operationId: createMenu
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysMenu" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/menus/{id}:
    put:
      tags: [RBAC]
      summary: 更新菜单
      operationId: updateMenu
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysMenu" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    delete:
      tags: [RBAC]
      summary: 删除菜单
      operationId: deleteMenu
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
...